added some changes after first review on 2 stage :)

+++ main-service/src/main/java/ru/practicum/mainservice/dto/event/EventFullDto.java
@@ -24,7 +24,7 @@ public class EventFullDto {
     private String annotation;
     @NotNull
     private CategoryDto category;
-    private int confirmedRequests;
+    private long confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @NotNull
     private LocalDateTime createdOn;
